batch rename breadcrumbs
