+++ enterprise/frontend/src/embedding-sdk/components/private/SdkBreadcrumbs/SdkBreadcrumbs.tsx
