+++ enterprise/frontend/src/embedding-sdk/components/private/SdkBreadcrumbs/SdkBreadcrumbsProvider.tsx
