+++ enterprise/frontend/src/embedding-sdk/components/private/SdkBreadcrumbs/index.ts
