+++ enterprise/frontend/src/embedding-sdk/components/private/SdkBreadcrumbs/utils/update-breadcrumbs.ts
